Modifies the directive data binding interface
It should fix issues #21 #24 #26
The main problem was on binding the video element
between the directive scope and the outside scope.
It was mainly caused because of the object reference
been directly overwritten inside of the directive.
Using an object to bind data fixes the issue,
because angular tracks changes in the properties
of that object, so the two-way data binding works.